Split rendering out into a separate crate
